+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 1.0.0
+## 0.1.1
+- Move result types to result package.
+- Create session package.
+- Create token package.
+- Create usecase package.
 
+## 0.1.0+1
+- Update README.md.
+
+## 0.1.0
 - Initial version.